Fast Math Functions

QuantumStatistics.FastMath — Module

Provide a set of fast math functions

QuantumStatistics.FastMath.invsqrt — Method

invsqrt(x)

The Legendary Fast Inverse Square Root See the following links: wikipedia and thesis

QuantumStatistics.Yeppp — Module

Forked from Yeppp.jl Please go to this link for instructions

QuantumStatistics.Yeppp.add — Method

add(x::Array, y::Array)

Perform element wise addition of the two array x and y.

QuantumStatistics.Yeppp.cos! — Method

cos(x)

Returns element wise cos of x.

QuantumStatistics.Yeppp.cos — Method

cos(x)

Returns element wise cos of x.

QuantumStatistics.Yeppp.dot — Method

dot(x::Vector{Float32}, y::Vector{Float32})

Compute the dot product of x and y.

QuantumStatistics.Yeppp.dot — Method

dot(x::Vector{Float64}, y::Vector{Float64})

Compute the dot product of x and y.

QuantumStatistics.Yeppp.dot — Method

dot(x, y)

Compute the dot product of x and y, which are two vectors with the same dimension.

QuantumStatistics.Yeppp.evalpoly — Method

evalpoly(coeff, x)

Evaluates polynomial with double precision (64-bit) floating-point coefficients coeff on an array of double precision (64-bit) floating-point elements x.

QuantumStatistics.Yeppp.exp! — Method

exp(x)

Computes the element wise exponential of x inplace.

QuantumStatistics.Yeppp.exp — Method

exp(x)

Returns the element wise exponential of x.

QuantumStatistics.Yeppp.log! — Method

log!(x)

Computes the element wise natural logarithm of x inplace.

QuantumStatistics.Yeppp.log — Method

log(x)

Returns the element wise natural logarithm of x.

QuantumStatistics.Yeppp.max — Method

max(x, y)

Compares the vectors x and y and return the element wise maxima.

QuantumStatistics.Yeppp.min — Method

max(x, y)

Compares the vectors x and y and return the element wise minima.

QuantumStatistics.Yeppp.multiply — Method

multiply(x, y)

Perform element wise multiplication of the two array x and y.

QuantumStatistics.Yeppp.sin! — Method

sin(x)

Computes element wise sin of x inplace.

QuantumStatistics.Yeppp.sin — Method

sin(x)

Returns element wise sin of x.

QuantumStatistics.Yeppp.subtract — Method

subtract(x::Array, y::Array)

Perform element wise subtraction of the two array x and y.

QuantumStatistics.Yeppp.tan! — Method

tan(x)

Returns element wise tan of x.

QuantumStatistics.Yeppp.tan — Method

tan(x)

Returns element wise tan of x.