Green's functions

QuantumStatistics.Green — Module

Provide N-body response and correlation functions

QuantumStatistics.Green.bareFermi — Method

calcualte with a given momentum vector and the chemical potential μ, rotation symmetry is assumed.

QuantumStatistics.Green.bareFermi — Method

bareFermi(β, τ, ε, [, scale])

Compute the bare fermionic Green's function. Assume $k_B=\hbar=1$

\[g(τ>0) = e^{-ετ}/(1+e^{-βε}), g(τ≤0) = -e^{-ετ}/(1+e^{βε})\]

Arguments

β: the inverse temperature

τ: the imaginary time, must be (-β, β]

ε: dispersion minus chemical potential: $E_k-μ$ it could also be the real frequency ω if the bare Green's function is used as the kernel in the Lehmann representation

QuantumStatistics.Green.bareFermiMatsubara — Method

bareFermiMatsubara(β, n, ε, [, scale])

Compute the bare Green's function for a given Matsubara frequency.

\[g(iω_n) = -1/(iω_n-ε),\]

where $ω_n=(2n+1)π/β$. The convention here is consist with the book "Quantum Many-particle Systems" by J. Negele and H. Orland, Page 95

Arguments

β: the inverse temperature

τ: the imaginary time, must be (-β, β]

ε: dispersion minus chemical potential: $E_k-μ$; it could also be the real frequency ω if the bare Green's function is used as the kernel in the Lehmann representation

QuantumStatistics.Green.FermiDirac — Method

FermiDirac(β, ε)

Compute the Fermi Dirac function. Assume $k_B=\hbar=1$

\[f(ϵ) = 1/(1+e^{-βε})\]

Arguments

β: the inverse temperature

ε: dispersion minus chemical potential: $E_k-μ$ it could also be the real frequency ω if the bare Green's function is used as the kernel in the Lehmann representation