Grids

QuantumStatistics.Grid.Uniform — Type

Uniform{Type,SIZE}

Create a uniform Grid with a given type and size

Member:

β: inverse temperature

halfLife: the grid is densest in the range (0, halfLife) and (β-halfLife, β)

size: the Grid size

grid: vector stores the grid

size: size of the grid vector

head: grid head

tail: grid tail

δ: distance between two grid elements

isopen: if isopen[1]==true, then grid[1] will be slightly larger than the grid head. Same for the tail.

QuantumStatistics.Grid.boseK — Function

boseK(Kf, maxK, halfLife, size::Int, kFi = floor(Int, 0.5size), twokFi = floor(Int, 2size / 3), type = Float64)

Create a logarithmic bosonic K Grid, which is densest near the momentum 0 and 2k_F

#Arguments:

Kf: Fermi momentum

maxK: the upper bound of the grid

halfLife: the grid is densest in the range (0, Kf+halfLife) and (2Kf-halfLife, 2Kf+halfLife)

size: the Grid size

kFi: index of Kf

twokFi: index of 2Kf

QuantumStatistics.Grid.fermiK — Function

fermiK(Kf, maxK, halfLife, size::Int, kFi = floor(Int, 0.5size), type = Float64)

Create a logarithmic fermionic K Grid, which is densest near the Fermi momentum $k_F$

#Arguments:

Kf: Fermi momentum

maxK: the upper bound of the grid

halfLife: the grid is densest in the range (Kf-halfLife, Kf+halfLife)

size: the Grid size

kFi: index of Kf

QuantumStatistics.Grid.tau — Function

tau(β, halfLife, size::Int, type = Float64)

Create a logarithmic Grid for the imaginary time, which is densest near the 0 and β

#Arguments:

β: inverse temperature

halfLife: the grid is densest in the range (0, halfLife) and (β-halfLife, β)

size: the Grid size