Green's Function

GreenFunc.Green2DLR — Type

Green's function with two external legs that has in-built Discrete Lehmann Representation. #Parameters:

'T': type of data

'TType': type of time domain, TType<:TimeDomain

'TGT': type of time grid

'SGT': type of space grid

#Members:

'name': Name of green's function

'color': Number of different species of Green's function (such as different spin values)

'dlrGrid': In-built Discrete Lehmann Representation

'timeGrid': Time or Frequency grid

'spaceType': Whether the Green's function is in coordinate space/momentum space

'spaceGrid': Coordinate or momentum grid

'instant': Instantaneous part of Green's function that is proportional to δ(τ) in τ space.

'dynamic': Dynamic part of Green's function

'instantError': Error of instantaneous part

'dynamicError': Error of dynamic part

GreenFunc.dynamic — Method

function dynamic(green::Green2DLR{DT,TT,TGT,SGT}, time, space, color1::Int, color2::Int, timeMethod::TM , spaceMethod::SM) where {DT,TT,TGT<:CompositeGrids.AbstractGrid,SGT<:CompositeGrids.AbstractGrid,TM,SM}

Find value of Green's function's dynamic part at given color and k/x by interpolation. Interpolation method is by default depending on the grid, but could also be chosen to be linear.

#Argument

'green': Green's function

'time': Target τ/ω_n point

'space': Target k/x point

'color1': Target color1

'color2': Target color2

'timeMethod': Method of interpolation for time

'spaceMethod': Method of interpolation for space

GreenFunc.instant — Method

function instant(green::Green2DLR{DT,TT,TGT,SGT}, space, color1::Int, color2::Int=color1; spaceMethod::SM = DEFAULTINTERP) where {DT,TT,TGT,SGT,SM}

Find value of Green's function's instant part at given color and k/x by interpolation. Interpolation method is by default depending on the grid, but could also be chosen to be linear.

#Argument

'green': Green's function

'space': Target k/x point

'color1': Target color1

'color2': Target color2

'spaceMethod': Method of interpolation for space.

GreenFunc.toDLR — Method

function toDLR(green::Green2DLR)

Convert Green's function to dlr space.

#Arguements

'green': Original Green's function

GreenFunc.toMatFreq — Function

function toMatFreq(green::Green2DLR, targetGrid =  green.dlrGrid.n)

Convert Green's function to matfreq space by Fourier transform. If green is already in matfreq space then it will be interpolated to the new grid.

#Arguements

'green': Original Green's function

'targetGrid': Grid of outcome Green's function. Default: DLR n grid

GreenFunc.toTau — Function

function toTau(green::Green2DLR, targetGrid =  green.dlrGrid.τ)

Convert Green's function to τ space by Fourier transform. If green is already in τ space then it will be interpolated to the new grid.

#Arguements

'green': Original Green's function

'targetGrid': Grid of outcome Green's function. Default: DLR τ grid